Introducing superman, a talk-to ai agent
I find qwen2.5:0.5b to be the best opensource model out there due to small size & great quality.

However, the chinese origin can come off as hard to scale due to the censorship nature in it.

The next best option is llama3.2:1b, but I find it too big for a RPi.

Which should I pick?